% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bc_idp_pais_controlador}
\alias{bc_idp_pais_controlador}
\title{Investimento Direto no País, Controlador Final}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 590 rows and 3 columns.
}
\source{
\url{https://www.bcb.gov.br/estatisticas/tabelasespeciais}
}
\usage{
bc_idp_pais_controlador
}
\description{
Dados consolidados de estoque de IDP, disponibilizados pelo Banco Central, até o ano de 2019,
pelo critério do controlador final.
}
\keyword{datasets}
