% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bc_idp_pais_provisorio}
\alias{bc_idp_pais_provisorio}
\title{Investimento Direto no País (IDP), dados provisórios}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 820 rows and 3 columns.
}
\source{
\url{https://www.bcb.gov.br/estatisticas/tabelasespeciais}
}
\usage{
bc_idp_pais_provisorio
}
\description{
Fluxos provisórios de IDP, disponibilizados pelo Banco Central,
discriminados por país, até o ano de 2020.
}
\keyword{datasets}
