% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bc_idp_pais_setor_part_capital}
\alias{bc_idp_pais_setor_part_capital}
\title{Investimento Direto no País, Participação no capital por setor}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 455 rows and 4 columns.
}
\source{
\url{https://www.bcb.gov.br/estatisticas/tabelasespeciais}
}
\usage{
bc_idp_pais_setor_part_capital
}
\description{
Dados consolidados de estoque de IDP, disponibilizados pelo Banco Central, até o ano de 2019,
por setor, pelo critério da participação no capital.
}
\keyword{datasets}
