% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bc_idp_posicao_provisorio}
\alias{bc_idp_posicao_provisorio}
\title{Posição de Investimento Internacional, dados provisórios}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2250 rows and 3 columns.
}
\source{
\url{https://www.bcb.gov.br/estatisticas/tabelasespeciais}
}
\usage{
bc_idp_posicao_provisorio
}
\description{
Estoque de IDP, disponibilizados pelo Banco Central, até o ano de 2020.
}
\keyword{datasets}
