% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrega_dados.R
\name{carrega_dados}
\alias{carrega_dados}
\title{Carrega dados brutos}
\usage{
carrega_dados(
  dados_brutos = "~\\\\sripts_r\\\\opcoes\\\\data\\\\negociacao-2022.xlsx"
)
}
\arguments{
\item{dados_brutos}{Endereço com .xlsx orinal da B3}
}
\value{
df com operações - fracionário, campos data, e nomes colunas corrigidos
}
\description{
Carrega dados brutos
}
\examples{
df_dados_brutos <- carrega_dados('~/data/arquivo_dados.xlsx')
}
