% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrega_dados.R
\name{corrige_fracionario}
\alias{corrige_fracionario}
\title{Corrige campo codigo_negociacao}
\usage{
corrige_fracionario(codigo_negociacao)
}
\arguments{
\item{codigo_negociacao}{A vector}
}
\value{
codigo_negociacao
}
\description{
Corrige campo codigo_negociacao
}
\examples{
mutate(codigo_negociacao = corrige_fracionario(codigo_negociacao))
}
