% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criar_tabela.R
\name{criar_tabela}
\alias{criar_tabela}
\title{Criar Tabela}
\usage{
criar_tabela(tabela, names)
}
\arguments{
\item{tabela}{Data frame}

\item{names}{Lista com os nomes das linhas}
}
\value{
Tabela ordenada com os valores dos países selecionados, e os nomes das linhas.
}
\description{
Criar Tabela
}
