% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDP_Qtd_Invest.R
\name{IDP_Qtd_Invest}
\alias{IDP_Qtd_Invest}
\title{Quantidade de Investidores IDP}
\usage{
IDP_Qtd_Invest(nomePlan, ano)
}
\arguments{
\item{nomePlan}{Planilha}

\item{ano}{Anos}
}
\value{
Data frame com os dados referente a quantidade de investidores dos paises selecionados
}
\description{
Quantidade de Investidores IDP
}
