% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_excel.R
\name{ler_excel}
\alias{ler_excel}
\title{Ler Excel}
\usage{
ler_excel(planName, sht, skp)
}
\arguments{
\item{planName}{Planilha}

\item{sht}{Aba da planilha}

\item{skp}{skip}
}
\value{
um objeto da classe data.frame, tbl, ou tbl_df
}
\description{
Ler Excel
}
