% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_linha.R
\name{ler_linha}
\alias{ler_linha}
\title{Ler linha}
\usage{
ler_linha(dataframe, len)
}
\arguments{
\item{dataframe}{Data frame}

\item{len}{Quantidade de colunas}
}
\value{
Data frame filtrado com os países selecionados, colunas renomeadas, e colunas em type numeric
}
\description{
Ler linha
}
