% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outros_func.R
\name{outros_func}
\alias{outros_func}
\title{Outros}
\usage{
<<<<<<< HEAD
outros_func(df_por_setor, plan3_reference, ano_referente)
=======
outros_func(df_por_setor, plan3_reference)
>>>>>>> 0147e82ab9d281451b940cfc09452d24b5a1f15d
}
\arguments{
\item{df_por_setor}{Data frame que contém os dados do Setor de atividade econômica}

\item{plan3_reference}{Referência à celula (L3 OU L4) da aba 3 da planilha de Investimentos}
}
\value{
Célula 'outros' da tabela do Setor de atividade econômica
}
\description{
Outros
}
