% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renomear_disc.R
\name{renomear_disc}
\alias{renomear_disc}
\title{renomear Discriminação}
\usage{
renomear_disc(dataframe)
}
\arguments{
\item{dataframe}{Data frame}
}
\value{
Renomeia a coluna Discriminação por Pais
}
\description{
Renomeia coluna
}
