% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrega_dados.R
\name{renomeia_coluna}
\alias{renomeia_coluna}
\title{Renomeia nomes das colunas}
\usage{
renomeia_coluna(df)
}
\arguments{
\item{df}{Data frame com colunas originais}
}
\value{
df_nome_col
}
\description{
Renomeia nomes das colunas
}
\examples{
df_renomeado <- renomeia_coluna(df)
}
