% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecionarPais.R
\name{selecionarPais}
\alias{selecionarPais}
\title{selecionarPais}
\usage{
selecionarPais(dataFrame)
}
\arguments{
\item{dataFrame}{data frame}
}
\value{
um data frame filtrado por país e com os anos que foram selecionados
}
\description{
Países selecionados
}
