% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setores.R
\name{setores}
\alias{setores}
\title{setores}
\usage{
setores(nomePlan)
}
\arguments{
\item{nomePlan}{Planilha}
}
\value{
Data frame contendo os setores selecionados e os valores de cada país
}
\description{
setores
}
