% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setores_final.R
\name{setores_final}
\alias{setores_final}
\title{Setores Final}
\usage{
setores_final(nomePlan, placeholder)
}
\arguments{
\item{nomePlan}{Planilha}

\item{placeholder}{Segunda planilha para guardar os dados da primeira}
}
\value{
Data frame contendo os setores selecionados e os valores dos países somados transformando em apenas uma linha
}
\description{
Soma os valores
}
