% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setores_idb.R
\name{setores_idb}
\alias{setores_idb}
\title{Setores IDB}
\usage{
setores_idb(nomePlan)
}
\arguments{
\item{nomePlan}{Data frame}
}
\value{
Data frame contendo somente os setores e países selecionados para a criação da tabela de Setor de atividade econômica
}
\description{
Setores IDB
}
