% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soma_linhas.R
\name{soma_linhas}
\alias{soma_linhas}
\title{Soma de linhas de cada país}
\usage{
soma_linhas(nlinha)
}
\arguments{
\item{nlinha}{o nome de um dataframe que contenha dados dos países de 2010 a 2020}
}
\value{
Soma das linhas, transformando em uma linha apenas
}
\description{
Soma de linhas de cada país
}
