% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soma_Qtd_Invest.R
\name{soma_Qtd_Invest}
\alias{soma_Qtd_Invest}
\title{Soma de Quantidade de Investidores}
\usage{
soma_Qtd_Invest(nlinha, len, ano)
}
\arguments{
\item{nlinha}{Data frame}

\item{len}{Quantidade de colunas}

\item{ano}{Anos}
}
\value{
Data frame contendo o somatório da quantidade de investidores dos países
}
\description{
Soma de Quantidade de Investidores
}
