% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soma_Qtd_Invest_IDP.R
\name{soma_Qtd_Invest_IDP}
\alias{soma_Qtd_Invest_IDP}
\title{Soma de Quantidade de Investidores}
\usage{
soma_Qtd_Invest_IDP(nlinha, len, anos)
}
\arguments{
\item{nlinha}{Data frame}

\item{len}{Quantidade de colunas}

\item{anos}{Anos}
}
\value{
Data frame contendo o somatório da quantidade de investidores dos países
}
\description{
Soma de Quantidade de Investidores
}
