% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turn_numeric.R
\name{turn_numeric}
\alias{turn_numeric}
\title{turn numeric}
\usage{
turn_numeric(dataframe, len)
}
\arguments{
\item{dataframe}{Data frame}

\item{len}{quantidade de colunas}
}
\value{
Colunas no type numeric
}
\description{
turn numeric
}
