% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unctad_wir}
\alias{unctad_wir}
\title{Dados do World Investment Report 2020, UNCTAD}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 60382 rows and 17 columns.
}
\source{
\url{https://unctadstat.unctad.org}
}
\usage{
unctad_wir
}
\description{
Dados do World Investment Report 2020, UNCTAD
}
\keyword{datasets}
